Final Upload
This contains all functionality of the FYP including UI, Machine learning pipelines, Lambda functions and file generation scripts - please note some biobert model files are omitted due to file size restrcitions

+++ UI/app.js
@@ -4,6 +4,9 @@ const CLIENT_ID = '51jji6cvtkpmqvqh3kf8m90b09';
 const REDIRECT_URI = 'http://localhost:5500';
 const DASHBOARD_URL = 'https://s6gly9n709.execute-api.eu-west-1.amazonaws.com/dev/dashboard';
 
+
+let allFiles =[];
+
 // Login function this redirects to Cognito hosted UI
 function login() {
     const url = `${COGNITO_DOMAIN}/login?client_id=${CLIENT_ID}&response_type=token&scope=email+openid+profile&redirect_uri=${encodeURIComponent(REDIRECT_URI)}`;
@@ -19,22 +22,26 @@ function logout() {
 
 // Check if user is logged in on page load
 window.onload = async () => {
-    // Check for token in URL hash
     const hash = window.location.hash.substring(1);
     const params = new URLSearchParams(hash);
-    const token = params.get('id_token');
-    
-    if (token) {
-        localStorage.setItem('id_token', token);
-        window.location.hash = ''; // Clean URL
+    const idToken = params.get('id_token');
+    const accessToken = params.get('access_token');
+
+    if (idToken) {
+        localStorage.setItem('id_token', idToken);
+    }
+    if (accessToken) {
+        localStorage.setItem('access_token', accessToken);
+    }
+    if (idToken || accessToken) {
+        window.location.hash = '';
     }
 
     const storedToken = localStorage.getItem('id_token');
-    
+
     if (storedToken) {
         document.getElementById('loginSection').style.display = 'none';
         document.getElementById('appSection').style.display = 'block';
-        
         const payload = JSON.parse(atob(storedToken.split('.')[1]));
         document.getElementById('welcomeMessage').textContent = `Welcome, ${payload.email}`;
     } else {
@@ -44,28 +51,31 @@ window.onload = async () => {
 };
 
 async function uploadFile() {
+
+    const token = localStorage.getItem('id_token');
     const fileInput = document.getElementById('fileInput');
     const status = document.getElementById('status');
     const file = fileInput.files[0];
 
     if (!file) {
-        status.textContent = 'Please select a file first!';
+        status.textContent = 'Please select a file first';
         return;
     }
 
     status.textContent = 'Uploading...';
 
-    // Convert file to base64
+    //Convert file to base64
     const reader = new FileReader();
     reader.readAsDataURL(file);
     reader.onload = async () => {
         const base64File = reader.result.split(',')[1];
 
-        try { 
+        try {
             const response = await fetch('https://s6gly9n709.execute-api.eu-west-1.amazonaws.com/dev/upload', {
                 method: 'POST',
                 headers: {
-                    'Content-Type': 'application/json'
+                    'Content-Type': 'application/json',
+                    'Authorization': token || ''
                 },
                 body: JSON.stringify({
                     fileName: file.name,
@@ -88,28 +98,62 @@ async function uploadFile() {
 
 
 async function listFiles() {
+    const token = localStorage.getItem('id_token');
+    console.log(token);
     const fileList = document.getElementById('fileList');
     fileList.innerHTML = '<tr><td colspan="3">Loading...</td></tr>';
 
     try {
         const response = await fetch('https://s6gly9n709.execute-api.eu-west-1.amazonaws.com/dev/listfiles', {
-            method: 'GET'
+            method: 'GET',
+            headers: {
+                'Authorization': token || ''
+            }
         });
 
         const raw = await response.json();
         const files = JSON.parse(raw.body);
+        allFiles = files;   
 
         if (files.length === 0) {
             fileList.innerHTML = '<tr><td colspan="3">No files in bucket</td></tr>';
             return;
         }
 
-        fileList.innerHTML = files.map(file => `
+        renderFileTable(files);
+
+    } catch (error) {
+        fileList.innerHTML = `<tr><td colspan="3"> Error: ${error.message}</td></tr>`;
+    }
+}
+
+function filterFiles(){
+    const search = document.getElementById('searchInput').value.toLowerCase();
+    const filtered = allFiles.filter(f => f.fileName.toLowerCase().includes(search));
+    renderFileTable(filtered);
+}
+
+function renderFileTable(files){
+    const fileList = document.getElementById('fileList');
+
+
+    const filtered = files.filter(f => !f.fileName.startsWith('model/'))
+    if(filtered.length === 0){
+
+        fileList.innerHTML = '<tr> <td> No files found</td></tr>';
+        return;
+    }
+
+    fileList.innerHTML = filtered.map(file => `
     <tr>
         <td>${file.fileName}</td>
         <td>${file.size}</td>
         <td>${file.lastModified}</td>
         <td>
+            <button onclick="downloadFile('${file.fileName}')" 
+                style="background: #001aff; color: white; border: none; padding: 5px 10px; border-radius: 4px; cursor: pointer;">
+                Download
+            </button>
             <button onclick="deleteFile('${file.fileName}')" 
                 style="background: #ff1900; color: white; border: none; padding: 5px 10px; border-radius: 4px; cursor: pointer;">
                 Delete
@@ -118,20 +162,57 @@ async function listFiles() {
     </tr>
 `).join('');
 
+}
+
+async function downloadFile(fileName) {
+    const token = localStorage.getItem('id_token');
+    const status = document.getElementById('status');
+    status.textContent = `Requesting access for ${fileName}...`;
+
+    try {
+        const response = await fetch(`https://s6gly9n709.execute-api.eu-west-1.amazonaws.com/dev/access?fileKey=${encodeURIComponent(fileName)}`, {
+            method: 'GET',
+            headers: { 'Authorization': token || '' }
+        });
+
+        const result = await response.json();
+
+        if (response.ok && result.downloadUrl) {
+            status.textContent = `Access granted, Downloading`;
+            const a = document.createElement('a');
+            a.href = result.downloadUrl;
+            a.download = fileName.split('/').pop();
+            document.body.appendChild(a);
+            a.click();
+            document.body.removeChild(a);
+        } else {
+            status.textContent = `${result.message || 'Access denied'}`;
+        }
     } catch (error) {
-        fileList.innerHTML = `<tr><td colspan="3">❌ Error: ${error.message}</td></tr>`;
+        status.textContent = `An error occured: ${error.message}`;
     }
 }
 
-
 async function loadDashboard() {
+
+    const token = localStorage.getItem('id_token');
+    console.log('Token being sent:', token ? token.substring(0, 50) : 'NULL');
     const status = document.getElementById('status');
     status.textContent = 'Loading dashboard...';
 
     try {
-        const response = await fetch(DASHBOARD_URL, { method: 'GET' });
-        const raw = await response.json();
-        const data = JSON.parse(raw.body);
+        const response = await fetch(DASHBOARD_URL, {
+            method: 'GET',
+            headers: {
+                'Authorization': token || ''
+            }
+        });
+        console.log('Status:', response.status);
+        const text = await response.text();
+        console.log('Raw response:', text);
+
+        const raw = JSON.parse(text);
+        const data = typeof raw.body === 'string' ? JSON.parse(raw.body) : raw;
 
         //Update summary cards
         document.getElementById('total').textContent = data.summary.total;
@@ -155,6 +236,18 @@ async function loadDashboard() {
                     <td><span class="badge ${getModelBadge(item.classifiedBy)}">${item.classifiedBy || '-'}</span></td>
                     <td>${item.uploadTimestamp || '-'}</td>
                     <td>${item.status || '-'}</td>
+                    <td>
+                        <select id="select-${item.ClassificationId}">
+                            <option value="RESTRICTED">RESTRICTED</option>
+                            <option value="INTERNAL">INTERNAL</option>
+                            <option value="PUBLIC">PUBLIC</option>
+                            <option value="QUARANTINE">QUARANTINE</option>
+                        </select>
+                        <button onclick="reclassifyFile('${item.fileKey}', '${item.ClassificationId}')"
+                            style="background: #2c3e50; color: white; border: none; padding: 5px 10px; border-radius: 4px; cursor: pointer; margin-left: 5px;">
+                            Reclassify
+                        </button>
+                    </td>
                 </tr>
             `).join('');
         }
@@ -199,6 +292,7 @@ function getModelBadge(model) {
 async function deleteFile(fileName) {
     if (!confirm(`Are you sure you want to permanently delete ${fileName}? This cannot be undone.`)) return;
 
+    const token = localStorage.getItem('id_token');
     const status = document.getElementById('status');
     status.textContent = `Deleting ${fileName}...`;
 
@@ -206,7 +300,8 @@ async function deleteFile(fileName) {
         const response = await fetch(`https://s6gly9n709.execute-api.eu-west-1.amazonaws.com/dev/delete?fileKey=${encodeURIComponent(fileName)}`, {
             method: 'DELETE',
             headers: {
-                'Content-Type': 'application/json'
+                'Content-Type': 'application/json',
+                'Authorization': token || ''
             }
         });
 
@@ -221,5 +316,41 @@ async function deleteFile(fileName) {
         }
     } catch (error) {
         status.textContent = `Delete failed: ${error.message}`;
+    }
+}
+
+
+async function reclassifyFile(fileKey, classificationId) {
+    const token = localStorage.getItem('id_token');
+    const newClassification = document.getElementById(`select-${classificationId}`).value;
+    const reason = prompt('Enter reason for reclassification:');
+
+    if (!reason) return;
+
+    try {
+        const response = await fetch('https://s6gly9n709.execute-api.eu-west-1.amazonaws.com/dev/reclassify', {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json',
+                'Authorization': token || ''
+            },
+            body: JSON.stringify({
+                fileKey: fileKey,
+                newClassification: newClassification,
+                reason: reason
+            })
+        });
+
+        const raw = await response.json();
+        const result = typeof raw.body === 'string' ? JSON.parse(raw.body) : raw;
+
+        if (response.ok) {
+            alert(`✅ ${result.message}`);
+            loadDashboard();
+        } else {
+            alert(`❌ ${result.message || 'Reclassification failed'}`);
+        }
+    } catch (error) {
+        alert(`❌ Error: ${error.message}`);
     }
 }
